write Category Controller & Service

+++ BulletinBoard.API/Services/CategoryService.cs
@@ -13,15 +13,13 @@ namespace BulletinBoard.API.Services
             {
                 using var db = new DataBaseContext();
                 var categories = db.Categories.ToList();
+                var subCategory = db.SubCategories.ToList();
                 var response = new List<GetAllCategoryResponse>();
+
                 categories.ForEach(parent =>
                 {
-                    var child = new List<SubCategoryModel>();
-                    parent.SubCategories.ForEach(sub =>
-                    {
-                        child.Add(new SubCategoryModel(sub.Id, sub.Name, sub.ImageUrl));
-                    });
-                    response.Add(new GetAllCategoryResponse(parent.Id, parent.Name, parent.ImageUrl, child));
+                    var subs = subCategory.Where(item => item.CategoryId == parent.Id).ToList();
+                    response.Add(new GetAllCategoryResponse(parent.Id, parent.Name, parent.ImageUrl, subs.ConvertAll(x => new SubCategoryModel(x.Id, x.Name, x.ImageUrl))));
                 });
 
                 return new BaseResponse<List<GetAllCategoryResponse>>(response);
@@ -40,10 +38,12 @@ namespace BulletinBoard.API.Services
                 if (request.ParentId == null)
                 {
                     db.Categories.Add(new Category(request.Name, request.ImageUrl));
+                    db.SaveChanges();
                 }
                 else
                 {
                     db.SubCategories.Add(new SubCategory(request.Name, request.ImageUrl, (int)request.ParentId));
+                    db.SaveChanges();
                 }
                 return new BaseResponse(0);
             }
@@ -95,6 +95,10 @@ namespace BulletinBoard.API.Services
                 {
                     var existCategory = db.Categories.FirstOrDefault(item => item.Id == request.Id);
                     if (existCategory == null) return new BaseResponse(1, "Такой категории не существует");
+
+                    var isHaveSubCategory = db.SubCategories.FirstOrDefault(item => item.CategoryId == request.Id);
+                    if (isHaveSubCategory != null) return new BaseResponse(1, "У категории есть подкатегории, удаление невозможно.");
+
                     db.Categories.Remove(existCategory);
                     db.SaveChanges();
                 }
